feat: separate global functions file, fix broken imports
