+++ src/app/layout.tsx
@@ -4,11 +4,11 @@ import { ThemeProvider } from '@mui/material/styles';
 import { GoogleAnalytics } from '@next/third-parties/google';
 import { type Metadata } from 'next';
 import { Bai_Jamjuree, Inter } from 'next/font/google';
-import { SnackbarProvider } from 'src/utils/snackbar';
 import { CheckRefreshToken } from '@src/components/auth/CheckRefreshToken';
 import { RegisterModalProvider } from '@src/components/global/RegisterModalProvider';
 import { TRPCReactProvider } from '@src/trpc/react';
 import ClientLocalizationProvider from '@src/utils/localization';
+import { SnackbarProvider } from '@src/utils/snackbar';
 import theme from '@src/utils/theme';
 
 const inter = Inter({
