+++ src/components/events/EventRegisterButton.tsx
@@ -6,10 +6,10 @@ import { Button, Skeleton, Tooltip } from '@mui/material';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { useRef } from 'react';
-import { setSnackbar, SnackbarPresets } from 'src/utils/snackbar';
 import { useRegisterModal } from '@src/components/global/RegisterModalProvider';
 import { useTRPC } from '@src/trpc/react';
 import { authClient } from '@src/utils/auth-client';
+import { setSnackbar, SnackbarPresets } from '@src/utils/snackbar';
 import EventEditButton from './EventEditButton';
 
 type EventRegisterButtonProps = {
