+++ src/components/manage/MemberList/MemberList.tsx
@@ -11,7 +11,6 @@ import {
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { TRPCClientErrorLike } from '@trpc/client';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useSnackbar } from 'src/utils/snackbar';
 import z from 'zod';
 import Confirmation from '@src/components/Confirmation';
 import { AppRouter } from '@src/server/api/root';
@@ -22,6 +21,7 @@ import {
 } from '@src/server/db/models';
 import { useTRPC } from '@src/trpc/react';
 import { authClient } from '@src/utils/auth-client';
+import { useSnackbar } from '@src/utils/snackbar';
 import CustomFooter from './CustomFooter';
 import CustomToolbar from './CustomToolbar';
 import { MemberListContext, MemberListContextType } from './MemberListContext';
