+++ src/components/settings/forms/UserInfo.tsx
@@ -6,7 +6,6 @@ import SchoolOutlinedIcon from '@mui/icons-material/SchoolOutlined';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
-import { setSnackbar, SnackbarPresets } from 'src/utils/snackbar';
 import Panel from '@src/components/common/Panel';
 import { majors, minors } from '@src/constants/utdDegrees';
 import { SelectUserMetadataWithClubs } from '@src/server/db/models';
@@ -17,6 +16,7 @@ import {
   AccountSettingsSchema,
   accountSettingsSchema,
 } from '@src/utils/formSchemas';
+import { setSnackbar, SnackbarPresets } from '@src/utils/snackbar';
 
 type UserInfoProps = {
   user: SelectUserMetadataWithClubs;
