+++ src/utils/snackbar/provider.tsx
@@ -8,66 +8,14 @@ import Snackbar, { SnackbarCloseReason } from '@mui/material/Snackbar';
 import SnackbarContent from '@mui/material/SnackbarContent';
 import { ReactNode, useCallback, useEffect, useRef, useState } from 'react';
 import { SnackbarContext, SnackbarContextDefault } from './context';
+import { useAttachGlobalSnackbarFunctions } from './global';
 import SnackbarPresets, { SnackbarDefault } from './presets';
 import {
   closeSnackbarFn,
   setSnackbarFn,
   setSnackbarWithPresetFn,
-  SnackbarFunctions,
   SnackbarType,
 } from './types';
-
-const globalContextFunctionsDefault = () => {
-  console.warn('Global snackbar functions not initialized');
-};
-
-const globalContextFunctionsUnmounted = () => {
-  console.warn('Snackbar provider unmounted');
-};
-
-/**
- * Mutable global snackbar functions
- */
-const globalSnackbarFunctions: SnackbarFunctions = {
-  setSnackbar: globalContextFunctionsDefault,
-  closeSnackbar: globalContextFunctionsDefault,
-  setSnackbarWithPreset: globalContextFunctionsDefault,
-};
-
-/**
- * Sets the global snackbar to the options defined in the function argument.
- *
- * Does not require calling the {@linkcode useSnackbar} hook
- *
- * @example <caption>Basic usage</caption>
- * setSnackbar("Lorem ipsum dolor sit amet");
- */
-export const setSnackbar: setSnackbarFn = (...args) =>
-  globalSnackbarFunctions.setSnackbar(...args);
-
-/**
- * Sets the global snackbar to a preset, along with any additional preset template parameters.
- *
- * Does not require calling the {@linkcode useSnackbar} hook
- *
- * @example <caption>Basic usage (with preset template parameter)</caption>
- * setSnackbarWithPreset("savedName", "lorem ipsum dolor sit amet");
- */
-export const setSnackbarWithPreset: setSnackbarWithPresetFn = (
-  preset,
-  ...args
-) => globalSnackbarFunctions.setSnackbarWithPreset(preset, ...args);
-
-/**
- * Closes the global snackbar. Can filter for a particular snackbar ID.
- *
- * Does not require calling the {@linkcode useSnackbar} hook
- *
- * @example <caption>Basic usage (with snackbar ID filter)</caption>
- * closeSnackbar("loginMessage");
- */
-export const closeSnackbar: closeSnackbarFn = (...args) =>
-  globalSnackbarFunctions.closeSnackbar(...args);
 
 /**
  * Provider component that provides snackbar context to its children and mounts the actual global snackbar components.
@@ -140,17 +88,10 @@ export const SnackbarProvider = ({ children }: { children: ReactNode }) => {
   };
 
   // Sync global snackbar functions with local component functions
-  useEffect(() => {
-    globalSnackbarFunctions.setSnackbar = setSnackbar;
-    globalSnackbarFunctions.setSnackbarWithPreset = setSnackbarWithPreset;
-    globalSnackbarFunctions.closeSnackbar = closeSnackbar;
-
-    return () => {
-      globalSnackbarFunctions.setSnackbar = globalContextFunctionsUnmounted;
-      globalSnackbarFunctions.setSnackbarWithPreset =
-        globalContextFunctionsUnmounted;
-      globalSnackbarFunctions.closeSnackbar = globalContextFunctionsUnmounted;
-    };
+  useAttachGlobalSnackbarFunctions({
+    setSnackbar,
+    setSnackbarWithPreset,
+    closeSnackbar,
   });
 
   const handleClose = useCallback(
